+++ netlify.toml
@@ -1,4 +1,4 @@
 [build]
     command = "pnpm build" # the command you run to build this file
-    functions = "functions" 
+    functions = "src/backend" 
     publish = "dist"
