+++ src/backend/jira-auth-init.ts
@@ -1,7 +1,7 @@
 import "dotenv/config";
 import { Handler } from "@netlify/functions";
 import * as client from "openid-client";
-import { StatusCode } from "../src/Utils/statuscode";
+import { StatusCode } from "../Utils/statuscode";
 import cookie from "cookie";
 import { cookieNames, getOpenidClientConfig } from "./utils";
